refactor(async): move UnserializableArgument to exceptions.rb; stop sniffing _aj_ keys
- Move Axn::Async::UnserializableArgument into lib/axn/exceptions.rb (the documented
  shared exception hierarchy per AGENTS.md), out of the internal serialization file.
- Drop the recursive `_aj_*`-key detection from deserialize. It was too greedy: a no-
  ActiveJob deployment can legitimately enqueue fallback JSON like `meta: {"_aj_version" => 1}`,
  and sniffing those keys routed deserialize to ActiveJob::Arguments on a worker without
  ActiveJob → NameError on valid data. Now dispatch only on the unambiguous `_as_global_id`
  fallback marker (decodable without ActiveJob), else use the ActiveJob decoder only when
  ActiveJob is actually loaded.
